fix(file-snapshot): turnId 改用随机 ID（ts_xxxxxx），不用下标
核心改动（用户提出）：
- turn_id 从 u32（下标 0,1,2...）改成 String（随机 hex 如 "ts_a3f8b2"）
- 每轮 turn on_turn_start 时生成唯一 ID
- 不依赖全局计数器，不需要扫磁盘找 max+1
- 多次 run / 回滚后继续 / --continue 不会覆盖旧快照

改动文件:
- snapshot.rs: ToolSnapshot.turn_id / DirSnapshot.turn_id: u32 → String
  + load_all_tool_snapshots() / load_tool_snapshots_after() 替代 range 查询
  + 按 timestamp 排序（不再按数字 turn_id 排序）
- mod.rs: current_turn: Mutex<u32> → current_turn_id: Mutex<String>
  + gen_turn_id() 生成 6 位随机 hex
  + on_turn_start 生成新 ID
- ion_worker.rs: fromTurn/toTurn 从 u64 改 String（turnId）
  + 范围过滤改用 timestamp 比较
- agent_loop.rs: 删掉 resolve_global_turn_start（不需要全局递增）
  恢复 turn_index=0（agent loop 内部计数器，与快照无关）

测试: 465 passed, 0 failed

+++ src/agent/agent_loop.rs
@@ -474,6 +474,8 @@ impl Agent {
     pub async fn run(&mut self, prompt: impl Into<String>) -> AgentResult<()> {
         self.running = true;
         self.stopped.store(false, std::sync::atomic::Ordering::SeqCst);
+        // turn_index 是 agent loop 内部计数器（每次 run 从 0 开始，用于 max_turns 限制）
+        // 快照用独立的全局唯一 turnId（ts_xxxxxx），不依赖 turn_index
         self.turn_index = 0;
 
         // ── 生命周期顺序 (对齐 pi) ──
@@ -569,7 +571,7 @@ impl Agent {
             self.check_pause().await?;
 
             let turn_ctx = TurnContext {
-                turn_index: turn as u64,
+                turn_index: turn,
                 messages: vec![],
                 has_tool_calls: false,
                 stop_reason: None,
